raster-nodes: move node `gradient_map` to its own mod
