+++ node-graph/graster-nodes/src/lib.rs
@@ -7,5 +7,6 @@ pub mod curve;
 pub mod dehaze;
 pub mod filter;
 pub mod generate_curves;
+pub mod gradient_map;
 pub mod image_color_palette;
 pub mod std_nodes;
